requirements.txt
# Core data + utils
pandas
numpy
tqdm
requests

# Images
pillow

# File formats
pyarrow   # for saving parquet files

# ML + retrieval (to add when you start embeddings)
torch
torchvision
open-clip-torch
faiss-cpu
transformers

# FastAPI server
fastapi
uvicorn[standard]
python-multipart
slowapi  # Rate limiting
